Added cmds and utils files and moved functions from app.c to them.
